Factored ConformImage interface out of common.py.
