+++ mriqc/messages.py
@@ -22,6 +22,7 @@ QC_UPLOAD_COMPLETE = "QC metrics successfully uploaded."
 QC_UPLOAD_START = "MRIQC Web API: submitting to <{url}>"
 GROUP_REPORT_GENERATED = "Group-{modality} report generated: {path}"
 RUN_FINISHED = "MRIQC completed."
+SUSPICIOUS_DATA_TYPE = "Input image {in_file} has a suspicious data type: '{dtype}'"
 TSV_GENERATED = "Generated summary TSV table for {modality} data: {path}"
 VOXEL_SIZE_OK = "Voxel size is large enough."
 VOXEL_SIZE_SMALL = "One or more voxel dimensions (%f, %f, %f) are smaller than the requested voxel size (%f) - diff=(%f, %f, %f)"
